refactor(web): hoist shared provider-form helpers out of the settings domain
Addresses two P2 review comments on #38643.

**Hoist shared form helpers (P2).** `components/speech/*-provider-form.tsx`
sits at the top level so both settings and chat can render it, but it still
reached back into `domains/settings/ai/` for its helpers — which routes the
chat first-run path through settings internals via a shared layer, bypassing
the cross-domain guard rather than satisfying it. The helpers those forms need
are now genuinely shared:

  domains/settings/ai/use-draft-override.ts  -> hooks/use-draft-override.ts
  domains/settings/ai/provider-catalogs.ts   -> lib/provider-catalogs.ts
  domains/settings/ai/local-storage-keys.ts  -> utils/local-settings-keys.ts

`shared-ui.tsx` is split by audience rather than moved wholesale: the three
domain-agnostic form controls (`SaveButton`, `ResetButton`, `CredentialsGuide`)
become `components/service-form-controls.tsx`, while the settings-page card
chrome (`ServiceCard`, `ByoServiceCard`, `ModeToggle`, `ManagedServicesBanner`,
`DomainVerificationChip`) stays in the settings domain, where its only callers
are.

The speech forms now import exclusively from top-level modules.

`chat/voice/stt-api.ts` kept its own copies of `LS_STT_PROVIDER`,
`LS_STT_API_KEY_PREFIX`, and `MACOS_NATIVE_STT_PROVIDER_ID` with a comment
explaining that cross-domain constants had to stay duplicated. The hoist
removes that constraint, so the duplicates are deleted and the file imports
the shared definitions (AGENTS.md — Single Source of Truth).

**Comments describe the present (P2).** The docstrings added to the two speech
forms narrated their own extraction ("split out from", "top-level rather than
... because two domains render it now") instead of describing the current API.
Rewritten as present-tense ownership and usage guidance; same for the
contrastive phrasing in the first-run card, voice picker modal, and voice
setting row (AGENTS.md — Code Comments).

Cross-domain allowlist unchanged at 12 entries.
