+++ clients/web/src/domains/settings/ai/web-search-card.tsx
@@ -19,15 +19,21 @@ import { Dropdown } from "@vellumai/design-library/components/dropdown";
 import { Input } from "@vellumai/design-library/components/input";
 import { toast } from "@vellumai/design-library/components/toast";
 
-import { ResetButton, SaveButton, ServiceCard } from "@/domains/settings/ai/shared-ui";
-import { LS_WEB_SEARCH_MODE, LS_WEB_SEARCH_PROVIDER } from "@/domains/settings/ai/local-storage-keys";
+import {
+  ServiceCard,
+} from "@/domains/settings/ai/shared-ui";
+import {
+  ResetButton,
+  SaveButton,
+} from "@/components/service-form-controls";
+import { LS_WEB_SEARCH_MODE, LS_WEB_SEARCH_PROVIDER } from "@/utils/local-settings-keys";
 import { getWebSearchProviderKeyStorage, parseServiceMode } from "@/domains/settings/ai/utils";
 import type { ServiceMode } from "@/generated/daemon/types.gen";
 import { useProvisionProviderKey } from "@/domains/settings/ai/use-daemon-config";
 import { useActiveAssistantId } from "@/assistant/use-active-assistant-id";
 import { configGetOptions, configGetSetQueryData, useConfigPatchMutation } from "@/generated/daemon/@tanstack/react-query.gen";
 import { useQuery } from "@tanstack/react-query";
-import { useDraftOverride } from "@/domains/settings/ai/use-draft-override";
+import { useDraftOverride } from "@/hooks/use-draft-override";
 import { credentialPresenceQueryKey, useStoredCredentialPresence } from "@/domains/settings/ai/use-stored-credential-presence";
 
 export function WebSearchCard() {
